+++ src/main/java/org/broadinstitute/hellbender/utils/downsampling/LevelingDownsampler.java
@@ -24,7 +24,7 @@ import java.util.*;
 public final class LevelingDownsampler<T extends List<E>, E> extends Downsampler<T> {
     private final int minElementsPerStack;
 
-    private final int targetSize;
+    private final long targetSize;
 
     private List<T> groups;
 
@@ -39,7 +39,7 @@ public final class LevelingDownsampler<T extends List<E>, E> extends Downsampler
      *                   this value -- if it does, items are removed from Lists evenly until the total size
      *                   is <= this value
      */
-    public LevelingDownsampler(final int targetSize ) {
+    public LevelingDownsampler(final long targetSize ) {
         this(targetSize, 1);
     }
 
@@ -53,7 +53,7 @@ public final class LevelingDownsampler<T extends List<E>, E> extends Downsampler
      *                            if a stack has only 3 elements and minElementsPerStack is 3, no matter what
      *                            we'll not reduce this stack below 3.
      */
-    public LevelingDownsampler(final int targetSize, final int minElementsPerStack ) {
+    public LevelingDownsampler(final long targetSize, final int minElementsPerStack ) {
         Utils.validateArg( targetSize >= 0, "targetSize must be >= 0 but got " + targetSize);
         Utils.validateArg( minElementsPerStack >= 0, "minElementsPerStack must be >= 0 but got " + minElementsPerStack);
 
@@ -142,7 +142,7 @@ public final class LevelingDownsampler<T extends List<E>, E> extends Downsampler
 
         // We will try to remove exactly this many items, however we will refuse to allow any
         // one group to fall below size 1, and so might end up removing fewer items than this
-        int numItemsToRemove = totalSize - targetSize;
+        long numItemsToRemove = totalSize - targetSize;
 
         currentGroupIndex = 0;
         int numConsecutiveUmodifiableGroups = 0;
